requirements.txt
streamlit
numpy
pandas
matplotlib
pandas_datareader
yfinance
keras.models
tensorflow
scikit-learn



